+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- [tanjun.clients.Client.from_gateway_bot][] can now also take cache-less `ShardAware` bots.
+
 ### Fixed
 - [tanjun.clients.Client.set_human_only][] no longer registers the internal check multiple
   times when called with [True][] multiple times.